Change stored data from JSON list to JSON dict
This model supports a dictionary of key-value pairs, where the key is
the term and the value is the list of classes per term.

The goal is to allow this application to scrape multiple terms at once.
Also, it will allow the application to store schedule information based
on the term.